test: extract OpenRouter white-box suite
