+++ scripts/rust-source-size-baseline.txt
@@ -35,7 +35,6 @@ crates/llm/src/chatgpt_responses.rs 1712
 crates/llm/src/google_gemini_generate_content.rs 1373
 crates/llm/src/lib.rs 1579
 crates/llm/src/openai_chat_completions.rs 3191
-crates/llm/src/openrouter.rs 1178
 crates/llmfs/src/lib.rs 1797
 crates/llmfs/tests/generation.rs 1627
 crates/service-manager/src/connection.rs 1048
